Change: Change episode values to float

+++ Episode.cs
@@ -3,7 +3,7 @@ class Episode
 {
     public string season { get; set; }
     public float episodeNumberOverall { get; set; } //EpisodeNumber
-    public int episodeNumberInSeason { get; set; } //EpisodeNumber2
+    public float episodeNumberInSeason { get; set; } //EpisodeNumber2
     public string title { get; set; }
     public string titleRomaji { get; set; }
     public string titleKanji { get; set; }
